Update LICENSE
This commit updates the project's license from GPL to BSD-3, providing more flexibility to users and contributors. The BSD-3 license allows for more permissive use of the project's code, without requiring derivative works to also be licensed under the same terms. This change will make the project more accessible to a wider range of users and increase opportunities for collaboration and innovation.