FIXED: Problems with chars that have more than one byte
Fixed a problem that occurs because of the method "len()" of the struct String, because it check's the bytes, not the chars, what's, in case of chars that have more then one byte, like special chars(ç, á, ã, í, •, ~ and etc.) or non-alphabetical characters. That's was causing an error of not giving the correct spacing and size for the entry name.