Update print.rs
Removed the non-ready function of printing the help command in a table, because the dynamic table is really on a early stage(and added the '[' and the ']' to the format function).
Changed the error function comment to something more relevant.
Added a "Edition" indicator instead of a "T" in the version function.
Using match instead of IF in the printer function to select how to print the entries. 
Changed the function format entries names to use the value entries instead of a reference, because it's is no more used after this function.